refactor: DebugRenderable.cs moved to a separate file
